% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_into_pvol.R
\name{dataframe_into_pvol}
\alias{dataframe_into_pvol}
\title{Add parameters to an existing pvol from a \code{data.frame}}
\usage{
dataframe_into_pvol(
  x,
  pvol,
  scan_id_column = "scan_id",
  range_column = "range",
  azimuth_column = "azimuth",
  to_add = "CLASS"
)
}
\arguments{
\item{x}{A data frame or tibble}

\item{pvol}{a polar volume as a \link[bioRad:summary.pvol]{pvol} object}

\item{to_add}{The name of the one or more columns to add}
}
\value{
A polar volume with the columns \code{to_add} from \code{x} have been added to.
}
\description{
Add parameters to an existing pvol from a \code{data.frame}
}
