% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Title}
\usage{
infer(x, model, n_features = NA)
}
\arguments{
\item{model}{Either a model or a math to a stored model}

\item{n_features}{The number of features if \code{model} is a path}
}
\value{
A vector with inferred class labels
}
\description{
Title
}
