% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvol_to_dataframe.R
\name{pvol_to_dataframe}
\alias{pvol_to_dataframe}
\title{Convert a pvol to a data frame}
\usage{
pvol_to_dataframe(
  pvol,
  to_retain = any_of(c("scan_id", "elevation", "azimuth", "range", "x", "y", "z", "DBZH",
    "DBZV", "TH", "TV", "VRADH", "VRADV", "WRADH", "WRADV", "PHIDP", "RHOHV", "KDP",
    "ZDR", "CLASS", "BIOLOGY"))
)
}
\arguments{
\item{to_retain}{}
}
\description{
Convert a pvol to a data frame
}
